feat: adiciona componente ClientRedirect para gerenciar redirecionamento baseado na linguagem do navegador
